% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTidyData.R
\name{createTidyData}
\alias{createTidyData}
\title{Create tidytext data structure}
\usage{
createTidyData(input_DF)
}
\arguments{
\item{data_df}{A data.frame containing the data}
}
\value{
A data.frame consisting turn, the tokenized dialogue, count and proportion of words within each turn.
\describe{
\item{turn}{Number indicating the turn of talk.}
\item{word}{The tokenized dialogue.}
\item{n}{The count of words per turn.}
\item{p}{The proportion of words per turn.}
}
}
\description{
This function transforms a data.frame having 3-columns of "person", "dialogue"
and "turn" into tidytext data structure. It tokenizes the dialogue, splitting each
sentence in separate words and remove stopwords. It also calculates count,
and proportion of word(s) within each turn.
}
\examples{
fict_df<-data.frame(person=c("John","Todd","Texas","Rama"),
dialogue=c("We gather here", "meet basic needs","people of this nation", "focus on implementation"),
turn=c(1,2,3,4), stringsAsFactors = FALSE)
createTidyData(fict_df)

}
\author{
Sewedo Abolade Todowede \email{todowede@gmail.com}
}
