% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDebateWords.R
\name{plotDebateWords}
\alias{plotDebateWords}
\title{Plot of change in proportion of given word(s) over time}
\usage{
plotDebateWords(word = NULL, data = debateData)
}
\arguments{
\item{word}{Single or combination of characters (words) for plotting}

\item{data}{A data.frame containing the data - Not required if supplied data is used.}
}
\value{
A ggplot graphic showing change in proportion of given word(s) over time.
}
\description{
This function creates a ggplot of change in proportion of given word(s)
over time, where time is indicated by turn
It calls the \code{createTidyData} function which prepares the data for plotting.
}
\examples{
fict_df<-data.frame(person=c("John","Todd","Texas","Rama"),
dialogue=c("We gather here", "meet basic needs","people of this nation", "focus on implementation"),
turn=c(1,2,3,4), stringsAsFactors = FALSE)
plotDebateWords(c("people","focus"),fict_df)

}
\author{
Sewedo Abolade Todowede \email{todowede@gmail.com}
}
