% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TidytextPlot-package.R
\docType{package}
\name{TidytextPlot-package}
\alias{TidytextPlot}
\alias{TidytextPlot-package}
\title{TidytextPlot: A Package to plot change in proportion of given words over time.}
\description{
The package transfrom a dataframe into a tidy text data structure and produces a plot showing how the proportion of given words changes over time.
}
\author{
\strong{Maintainer}: Sewedo Todowede \email{todowede@gmail.com}

}
\keyword{internal}
